Revert tpch to numeric

+++ schemas/tpch/sdb_all_tables_partition/schema.sql
@@ -39,7 +39,7 @@ CREATE TABLE part (
     p_type character varying(25) NOT NULL,
     p_size int NOT NULL,
     p_container character varying(10) NOT NULL,
-    p_retailprice double precision NOT NULL,
+    p_retailprice numeric(13,2) NOT NULL,
     p_comment character varying(23) NOT NULL
 ) PARTITION BY HASH (p_partkey);
 
@@ -59,7 +59,7 @@ CREATE TABLE supplier (
     s_address character varying(40) NOT NULL,
     s_nationkey int NOT NULL,
     s_phone character varying(15) NOT NULL,
-    s_acctbal double precision NOT NULL,
+    s_acctbal numeric(13,2) NOT NULL,
     s_comment character varying(101) NOT NULL
 ) PARTITION BY HASH (s_suppkey);
 
@@ -77,7 +77,7 @@ CREATE TABLE partsupp (
     ps_partkey int NOT NULL,
     ps_suppkey int NOT NULL,
     ps_availqty int NOT NULL,
-    ps_supplycost double precision NOT NULL,
+    ps_supplycost numeric(13,2) NOT NULL,
     ps_comment character varying(199) NOT NULL
 ) PARTITION BY HASH (ps_partkey);
 
@@ -97,7 +97,7 @@ CREATE TABLE customer (
     c_address character varying(40) NOT NULL,
     c_nationkey int NOT NULL,
     c_phone character varying(15) NOT NULL,
-    c_acctbal double precision NOT NULL,
+    c_acctbal numeric(13,2) NOT NULL,
     c_mktsegment character varying(10) NOT NULL,
     c_comment character varying(117) NOT NULL
 ) PARTITION BY HASH (c_custkey);
@@ -116,7 +116,7 @@ CREATE TABLE orders (
     o_orderkey bigint NOT NULL,
     o_custkey int NOT NULL,
     o_orderstatus character varying(1) NOT NULL,
-    o_totalprice double precision NOT NULL,
+    o_totalprice numeric(13,2) NOT NULL,
     o_orderdate date NOT NULL,
     o_orderpriority character varying(15) NOT NULL,
     o_clerk character varying(15) NOT NULL,
@@ -139,10 +139,10 @@ CREATE TABLE lineitem (
     l_partkey int NOT NULL,
     l_suppkey int NOT NULL,
     l_linenumber int NOT NULL,
-    l_quantity double precision NOT NULL,
-    l_extendedprice double precision NOT NULL,
-    l_discount double precision NOT NULL,
-    l_tax double precision NOT NULL,
+    l_quantity numeric(13,2) NOT NULL,
+    l_extendedprice numeric(13,2) NOT NULL,
+    l_discount numeric(13,2) NOT NULL,
+    l_tax numeric(13,2) NOT NULL,
     l_returnflag character varying(1) NOT NULL,
     l_linestatus character varying(1) NOT NULL,
     l_shipdate date NOT NULL,
